refactor(core): split session generate node
